# openapi.yaml
openapi: 3.0.3
info:
  title: Liquor Distribution OS API
  version: 1.0.0
  description: |
    REST API for the Liquor Distribution Operating System.
    Covers customers, orders, invoices, collections, deliveries, inventory, and sync.
    All mutating endpoints require an `Idempotency-Key` header.
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://staging.api.example.com/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers (route/risk/search filters)
      parameters:
        - { in: query, name: route_id, schema: { type: string } }
        - { in: query, name: risk, schema: { type: string, enum: [low, medium, high] } }
        - { in: query, name: status, schema: { type: string, enum: [active, hold, blocked, dispute] } }
        - { in: query, name: q, schema: { type: string }, description: name/code search }
        - { in: query, name: limit, schema: { type: integer, default: 50 } }
        - { in: query, name: cursor, schema: { type: string } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Customer' }
                  next_cursor: { type: string, nullable: true }
    post:
      tags: [Customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerInput' }
      responses:
        '201': { description: created }
  /customers/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Customers]
      summary: Get customer
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
    patch:
      tags: [Customers]
      summary: Update customer (non-credit fields)
      responses: { '200': { description: ok } }
  /customers/{id}/360:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Customers]
      summary: Full 360 view — profile, AR, history, suggestions
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer360' }
  /customers/{id}/credit:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    patch:
      tags: [Customers]
      summary: Update credit limit / payment term (admin+)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                credit_limit: { type: number }
                payment_term_id: { type: string }
                reason: { type: string }
      responses: { '200': { description: ok } }
  /customers/{id}/hold:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      tags: [Customers]
      summary: Place customer on hold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
                until: { type: string, format: date, nullable: true }
      responses: { '200': { description: ok } }

  # ───────────────────────────── CATALOG
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Customer:
      type: object
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        type: { type: string, enum: [outlet, bar, hotel, retailer] }
        route_id: { type: string }
        geo: { $ref: '#/components/schemas/GeoPoint' }
        address: { type: string }
        phone: { type: string }
        whatsapp: { type: string }
        assigned_rep_id: { type: string }
        assigned_collector_id: { type: string }
        credit_limit: { type: number }
        payment_term: { type: string, enum: [cash, same_day, net_7, net_14, net_30, pdc, custom] }
        status: { type: string, enum: [active, hold, blocked, dispute] }
        credit_state: { $ref: '#/components/schemas/CreditState' }
    CustomerInput:
      type: object
      required: [code, name, type, route_id, payment_term]
      properties:
        code: { type: string }
        name: { type: string }
        type: { type: string }
        route_id: { type: string }
        address: { type: string }
        phone: { type: string }
        whatsapp: { type: string }
        geo: { $ref: '#/components/schemas/GeoPoint' }
        credit_limit: { type: number }
        payment_term: { type: string }
    Customer360:
      type: object
      properties:
        customer: { $ref: '#/components/schemas/Customer' }
        aging: { $ref: '#/components/schemas/AgingBuckets' }
        recent_orders: { type: array, items: { type: object } }
        recent_payments: { type: array, items: { type: object } }
        open_invoices: { type: array, items: { type: object } }
        suggestions:
          type: object
          properties:
            likely_basket:
              type: array
              items:
                type: object
                properties:
                  product_id: { type: string }
                  qty: { type: integer }
            due_for_reorder: { type: boolean }
    CreditState:
      type: object
      properties:
        outstanding_total: { type: number }
        available_credit: { type: number }
        aging: { $ref: '#/components/schemas/AgingBuckets' }
        risk_score: { type: number, minimum: 0, maximum: 1 }
        last_payment_at: { type: string, format: date-time }
        last_visit_at: { type: string, format: date-time }
        promise_amount: { type: number }
        promise_due_date: { type: string, format: date }
    AgingBuckets:
      type: object
      properties:
        b_0_7: { type: number }
        b_8_15: { type: number }
        b_16_30: { type: number }
        b_31_60: { type: number }
        b_60_plus: { type: number }
...
    GeoPoint:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number }
        lng: { type: number }
        accuracy_m: { type: number }
...